Split regression test stories into their own file
